Improved hash function (incompatible algorithm change)
I've added a salt to the hash function to improve behavior when
handling similar byte values. For example, zero bytes aligned to
the size of a hash_t (64 bits) in the data block would not affect
the hash value at all. The salted version of the algorithm will
scramble the data with a preset value and ensure that input data
will almost always result in a different hash value for each
element introduced.

The algorithm was further optimized by only dereferencing the
data pointer once instead of at every operation.

Testing this change on two hard drive image files with 4KB blocks,
the number of block hash collisions over the entire data set
dropped to zero for the entire ~9GB of data while performance did
not drop enough to register.