Make hash configurable for 64, 32, or 16 bit widths
In some instances it would be handy to produce hashes that are not
64 bits wide, especially on systems with smaller CPUs. I still
need to test whether the hashes are unique enough, but collision
risk always goes up with less bits anyway. The narrower versions
could be suitable for scenarios like hashing strings, where the
data is not very large in the first place.