v5.0: use better hash shift value
Testing a total of 14,612,062 text strings for collisions showed
a big improvement changing the hash shift from 11 to 14 bits. The
text includes all English dictionary words in upper/lower cases,
all numbers from 1 to 1000000, and a huge list of path names
(absolute and relative) generated using find and then cutting
down path compnents with sed and merging all resulting relative
path names.

Collisions reduced from 14434 hashes to 482 hashes in tests. The
new -l option for the jodyhash utility was used to perform these
tests quickly.

It should also be noted that 18, 22, 30, and 31 also showed good
collision results, with 30 and 31 coming very close to 14. In the
32-bit width tests, 6, 10, 14, 18, 22, and 26 were good but 30
and 31 were not.

In general, powers of 2 are the absolute worst shift values. 4 is
very bad in 64-bit width. 16 is terrible in 32-bit width and there
are several other values in 32-bit width that weren't much better.
The "good" values given previously are in fact the only values
for 32-bit width that do not have an abysmal collision rate.