Add 4K block hash mode, fix a bug, update copyright year
This change adds the -B option which will hash every 4K block of
the file in order and output the hash of that block. This can be
useful for locating identical data in different files.

A coding error was fixed where PRINTHASH() uses the wrong variable
in the #define. In practice, this variable was always 'hash', so
compiled code would still work, but this change makes it actually
correct rather than wrong but functional.